+++ lib/karafka/pro/processing/strategies/dlq_lrj.rb
@@ -46,7 +46,7 @@ module Karafka
                 skippable_message = find_skippable_message
 
                 unless revoked?
-                  copy_skippable_message_to_dlq(skippable_message)
+                  dispatch_skippable_message_to_dlq(skippable_message)
                   mark_as_consumed(skippable_message)
                 end