+++ spec/integrations/consumption/strategies/dlq_mom/multi_partition_target_flow_spec.rb
@@ -4,7 +4,6 @@
 
 setup_karafka(allow_errors: %w[consumer.consume.error])
 
-create_topic(name: DT.topics[0])
 create_topic(name: DT.topics[1], partitions: 10)
 
 class Consumer < Karafka::BaseConsumer