+++ spec/integrations/consumption/with_separate_subscription_groups_spec.rb
@@ -20,8 +20,6 @@ draw_routes do
         consumer Consumer
       end
     end
-
-    create_topic(name: topic_name)
   end
 end