+++ spec/integrations/pro/consumption/strategies/dlq/dispatch_instrumentation_spec.rb
@@ -1,13 +1,10 @@
 # frozen_string_literal: true
 
-# Same as pure DLQ version until rebalance
-# Needs to go to same partition
+# When DLQ delegation happens, Karafka should emit appropriate event.
 
 setup_karafka(allow_errors: %w[consumer.consume.error]) do |config|
   config.license.token = pro_license_token
 end
-
-create_topic(name: DT.topics[1], partitions: 10)
 
 class Consumer < Karafka::Pro::BaseConsumer
   def consume
@@ -15,32 +12,23 @@ class Consumer < Karafka::Pro::BaseConsumer
   end
 end
 
-class DlqConsumer < Karafka::Pro::BaseConsumer
-  def consume
-    messages.each do |message|
-      DT[:broken] << message.partition
-    end
-  end
-end
-
 draw_routes do
   topic DT.topics[0] do
     consumer Consumer
     dead_letter_queue(topic: DT.topics[1], max_retries: 0)
-    manual_offset_management true
   end
+end
 
-  topic DT.topics[1] do
-    consumer DlqConsumer
-    manual_offset_management true
-  end
+Karafka.monitor.subscribe('dead_letter_queue.dispatched') do |event|
+  assert !event[:message].nil?
+  DT[:events] << 1
 end
 
 elements = DT.uuids(100)
 produce_many(DT.topic, elements)
 
 start_karafka_and_wait_until do
-  DT[:broken].count >= 10
+  DT[:events].size.positive?
 end
 
-assert_equal 1, DT[:broken].uniq.count
+# No need for specs, if event was dispatched, we will stop, otherwise hangs.